+++ src/test/java/de/kleemann/co2_hsharz/api/emission/EmissionsCalculationControllerTests.java
@@ -13,9 +13,9 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 import de.kleemann.co2_hsharz.api.emission.EmissionsCalculationTestsUtil.TransportMediumDTOValidation;
 import de.kleemann.co2_hsharz.core.transport.TransportMediumService;
-import de.kleemann.co2_hsharz.persistence.fuel.TransportMediumFuel;
 import de.kleemann.co2_hsharz.persistence.transport.TransportMediumName;
-import de.kleemann.co2_hsharz.persistence.transport.TransportMediumSize;
+import de.kleemann.co2_hsharz.persistence.transport.fuel.TransportMediumFuel;
+import de.kleemann.co2_hsharz.persistence.transport.size.TransportMediumSize;
 
 /**
  * This class contains tests for {@link EmissionsCalculationController}
