+++ src/test/java/de/kleemann/co2_hsharz/persistence/transport/TransportMediumRepositoryTests.java
@@ -11,7 +11,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
-import de.kleemann.co2_hsharz.persistence.fuel.TransportMediumFuel;
+import de.kleemann.co2_hsharz.persistence.transport.fuel.TransportMediumFuel;
 
 /**
  * This class contains tests for {@link TransportMediumRepository}
