+++ src/test/java/de/kleemann/co2_hsharz/persistence/transport/TransportMediumTestUtil.java
@@ -1,6 +1,7 @@
 package de.kleemann.co2_hsharz.persistence.transport;
 
-import de.kleemann.co2_hsharz.persistence.fuel.TransportMediumFuel;
+import de.kleemann.co2_hsharz.persistence.transport.fuel.TransportMediumFuel;
+import de.kleemann.co2_hsharz.persistence.transport.size.TransportMediumSize;
 
 public class TransportMediumTestUtil {
